moving achilles result dist dao calls to service
