+++ public-api/db-cdr/csv/achilles_analysis.csv
@@ -1,0 +1,46 @@
+"analysis_id","analysis_name","stratum_1_name","stratum_2_name","stratum_3_name","stratum_4_name","stratum_5_name","stratum_6_name","stratum_7_name","chart_type","data_type"
+0,"Source name","","","","","",NULL,NULL,"column","counts"
+1,"Number of persons","","","","","",NULL,NULL,"column","counts"
+2,"Number of persons by sex at birth (ppi variable","gender_concept_id","","","","",NULL,NULL,"column","counts"
+3,"Number of persons by year of birth","year_of_birth","","","","",NULL,NULL,"column","counts"
+4,"Number of persons by race","race_concept_id","","","","",NULL,NULL,"pie","counts"
+5,"Number of persons by ethnicity","ethnicity_concept_id","","","","",NULL,NULL,"column","counts"
+6,"Number of persons by age decile","age_decile","","","","",NULL,NULL,"column","counts"
+7,"Number of persons by gender_identity","gender_identity_concept_id","","","","",NULL,NULL,"column","counts"
+8,"Number of persons by race/ethnicity a/c to ppi data","concept_id","concept_name","","","",NULL,NULL,"column","counts"
+10,"Number of all persons by year of birth by gender","year_of_birth","gender_concept_id","","","",NULL,NULL,"column","counts"
+12,"Number of persons by race and ethnicity","race_concept_id","ethnicity_concept_id","","","",NULL,NULL,"pie","counts"
+100,"Physical Measurements Concept Group Metadata","group_title","group_name","concept_1","concept_2","concept_3",NULL,NULL,NULL,NULL
+1814,"Distribution of numeric values by measurement concept id ","measurement_concept_id","unit_name",NULL,NULL,NULL,NULL,NULL,"box","distribution"
+1815,"Distribution of numeric values, by measurement_concept_id and gender","measurement_concept_id","unit_name","gender_concept_id","IQR_min","IQR_max",NULL,NULL,"box","distribution"
+1816,"Distribution of numeric values by measurement_concept_id and age at occurrence decile","measurement_concept_id","unit_name","age_decile","IQR_min","IQR_max",NULL,NULL,"box","distribution"
+1900,"Sex Assigned at Birth","measurement_concept_id","unit_name","gender_concept_id","value","value_as_concept_id",NULL,NULL,"column","counts"
+1901,"Measurement response age at occurrence histogram","measurement_concept_id","unit_name","age_decile","value","value_as_concept_id",NULL,NULL,"column","counts"
+1910,"Measurement response counts by unit and gender","measurement_concept_id","unit_name","gender_concept_id","","",NULL,NULL,"column","counts"
+3000,"Participant Count","concept_id","","domain_id","","",NULL,NULL,"column","counts"
+3005,"Counts per person distribution by concept id","measurement_concept_id",NULL,NULL,NULL,NULL,NULL,NULL,"box","distribution"
+3006,"Counts per person by gender distribution","measurement_concept_id","gender_concept_id",NULL,NULL,NULL,NULL,NULL,"box","distribution"
+3007,"Counts per person by age at occurrence","measurement_concept_id","age_decile_at_occurrence",NULL,NULL,NULL,NULL,NULL,"box","distribution"
+3008,"Counts per person distrubtion by age at occurrence and gender","measurement_concept_id","age_decile_at_occurrence","gender_concept_id",NULL,NULL,NULL,NULL,"box","distribution"
+3101,"Biological Sex","concept_id","ppi_sex_at_birth_concept_id","domain_id",NULL,NULL,NULL,NULL,"column","counts"
+3102,"Age at Occurrence","concept_id","age_decile","domain_id","null","",NULL,NULL,"column","counts"
+3103,"Race","concept_id","race_concept_id","domain_id",NULL,NULL,NULL,NULL,"column","counts"
+3104,"Ethnicity","concept_id","ethnicity_concept_id","domain_id",NULL,NULL,NULL,NULL,"column","counts"
+3106,"Current Age histogram","concept_id","current_age_decile","domain_id",NULL,NULL,NULL,NULL,"column","counts"
+3107,"Participant count by date","concept_name","date","","","","","","line","counts"
+3110,"Response Distribution","survey_concept_id","question_concept_id","answer_concept_id","answer_value_string","question_order","question_path","hasSubQuestions","column","counts"
+3111,"Sex Assigned at Birth","survey_concept_id","question_concept_id","answer_concept_id","answer_value_string","gender_concept_id","question_path","version_id","column","counts"
+3112,"Age at occurrence","survey_concept_id","question_concept_id","answer_concept_id","answer_value_string","age_decile","question_path","version_id","column","counts"
+3113,"Survey response counts by version","survey_concept_id","question_concept_id","answer_concept_id","answer_value_string","question_order","question_path","version_id","column","counts"
+3200,"Survey Module counts by gender","survey_concept_id","gender_concept_id",NULL,NULL,NULL,NULL,NULL,"column","counts"
+3201,"Survey Module counts by age","survey_concept_id","age_decile",NULL,NULL,NULL,NULL,NULL,"column","counts"
+3300,"Domain counts by gender","domain_concept_id",NULL,"domain_name","gender_concept_id",NULL,NULL,NULL,"column","counts"
+3301,"Domain counts by age","domain_concept_id",NULL,"domain_name","gender_concept_id",NULL,NULL,NULL,"column","counts"
+3310,"Biological Sex","concept_id","ppi_sex_at_birth_concept_id","domain_id","count_percentage",NULL,NULL,"","column","percentage"
+3311,"Age at Occurrence","concept_id","age_decile","domain_id","count_percentage","",NULL,"","column","percentage"
+3320,"Sex Assigned at Birth","survey_concept_id","question_concept_id","","","gender_concept_id","question_path","","column","percentage"
+3321,"Age at occurrence","survey_concept_id","question_concept_id","","","age_decile","question_path",NULL,"column","percentage"
+3331,"Sex Assigned at Birth","survey_concept_id","question_concept_id","answer_concept_id","answer_value_string","gender_concept_id","question_path",NULL,"column","percentage_counts_by_question_counts"
+3332,"Age at occurrence","survey_concept_id","question_concept_id","answer_concept_id","answer_value_string","age_decile","question_path",NULL,"column","percentage_counts_by_question_counts"
+3400,"Survey Module Counts By Version","survey_concept_id","version_id","version_month",NULL,NULL,NULL,NULL,"",""
+3401,"Survey Question Counts By Version","survey_concept_id","version_id","version_month",NULL,NULL,NULL,NULL,"",""
