+++ public-api/db-cdr/csv/cb_criteria_physical_measurement.csv
@@ -1,0 +1,19 @@
+"id","parent_id","domain_id","is_standard","type","subtype","concept_id","code","name","value","est_count","is_group","is_selectable","has_attribute","has_hierarchy","has_ancestor_data","path","synonyms"
+325190,0,"PHYSICAL_MEASUREMENT",0,"PPI","BP","","","Blood Pressure","",-1,1,0,0,1,0,"","Blood Pressure"
+325191,325190,"PHYSICAL_MEASUREMENT",0,"PPI","BP","","","Hypotensive (Systolic <= 90 / Diastolic <= 60)","",10790,0,1,0,1,0,"","Hypotensive (Systolic <= 90 / Diastolic <= 60)|[physical_measurement_rank1]"
+325192,325190,"PHYSICAL_MEASUREMENT",0,"PPI","BP","","","Normal (Systolic <= 120 / Diastolic <= 80)","",63366,0,1,0,1,0,"","Normal (Systolic <= 120 / Diastolic <= 80)|[physical_measurement_rank1]"
+325193,325190,"PHYSICAL_MEASUREMENT",0,"PPI","BP","","","Pre-Hypertensive (Systolic 121-139 / Diastolic 81-89)","",6149,0,1,0,1,0,"","Pre-Hypertensive (Systolic 121-139 / Diastolic 81-89)|[physical_measurement_rank1]"
+325194,325190,"PHYSICAL_MEASUREMENT",0,"PPI","BP","","","Hypertensive (Systolic >= 140 / Diastolic >= 90)","",94090,0,1,0,1,0,"","Hypertensive (Systolic >= 140 / Diastolic >= 90)|[physical_measurement_rank1]"
+325195,325190,"PHYSICAL_MEASUREMENT",0,"PPI","BP","","","Blood Pressure","",0,0,1,1,1,0,"","Blood Pressure|[physical_measurement_rank1]"
+325196,0,"PHYSICAL_MEASUREMENT",0,"PPI","HR","","","Heart Rate","",-1,1,0,0,1,0,"","Heart Rate"
+325197,325196,"PHYSICAL_MEASUREMENT",0,"PPI","HR","","","Heart Rhythm Status","",-1,1,0,0,1,0,"","Heart Rhythm Status"
+325198,325197,"PHYSICAL_MEASUREMENT",0,"PPI","HR-IRR",1586218,"","Irregularity detected",4262985,0,0,1,0,1,0,"","Irregularity detected|heart-rhythm-status|heart-rhythm-status|[physical_measurement_rank1]"
+325199,325197,"PHYSICAL_MEASUREMENT",0,"PPI","HR-NOIRR",1586218,"","No-irregularity detected",4297303,0,0,1,0,1,0,"","No-irregularity detected|heart-rhythm-status|heart-rhythm-status|[physical_measurement_rank1]"
+325207,0,"PHYSICAL_MEASUREMENT",0,"PPI","WHEEL",903111,"","Wheelchair user at enrollment",4023190,0,0,1,0,1,0,"","Wheelchair user at enrollment|Participant is a wheelchair user|Participant is a wheelchair user|[physical_measurement_rank1]"
+325206,0,"PHYSICAL_MEASUREMENT",0,"PPI","PREG",903120,"","Pregnant at enrollment",45877994,0,0,1,0,1,0,"","Pregnant at enrollment|Participant is pregnant|Participant is pregnant|[physical_measurement_rank1]"
+325202,0,"PHYSICAL_MEASUREMENT",0,"PPI","WEIGHT",903121,"","Weight","",408435,0,1,1,1,0,"","Weight|Weight|Weight|[physical_measurement_rank1]"
+325203,0,"PHYSICAL_MEASUREMENT",0,"PPI","BMI",903124,"","BMI","",408473,0,1,1,1,0,"","BMI|Computed body mass index|Computed body mass index|[physical_measurement_rank1]"
+325200,325196,"PHYSICAL_MEASUREMENT",0,"PPI","HR-DETAIL",903126,"","Heart Rate","",408339,0,1,1,1,0,"","Heart Rate|Computed heart rate, mean of 2nd and 3rd measures|Computed heart rate, mean of 2nd and 3rd measures|[physical_measurement_rank1]"
+325201,0,"PHYSICAL_MEASUREMENT",0,"PPI","HEIGHT",903133,"","Height","",408041,0,1,1,1,0,"","Height|Height|Height|[physical_measurement_rank1]"
+325204,0,"PHYSICAL_MEASUREMENT",0,"PPI","WC",903135,"","Waist Circumference","",408407,0,1,1,1,0,"","Waist Circumference|Computed waist circumference, mean of closest two measures|Computed waist circumference, mean of closest two measures|[physical_measurement_rank1]"
+325205,0,"PHYSICAL_MEASUREMENT",0,"PPI","HC",903136,"","Hip Circumference","",408047,0,1,1,1,0,"","Hip Circumference|Computed hip circumference, mean of closest two measures|Computed hip circumference, mean of closest two measures|[physical_measurement_rank1]"
