+++ public-api/db-cdr/csv/domain_info.csv
@@ -1,0 +1,7 @@
+"concept_id","domain","domain_id","name","description","all_concept_count","standard_concept_count","participant_count"
+19,0,"Condition","Conditions","Conditions are records of a Person suggesting the presence of a disease or medical condition stated as a diagnosis, a sign or a symptom, which is either observed by a Provider or reported by the patient.",73850,24773,192000
+13,3,"Drug","Drug Exposures","Drugs biochemical substance formulated in such a way that when administered to a Person it will exert a certain physiological or biochemical effect. The drug exposure domain concepts capture records about the utilization of a Drug when ingested or otherwise introduced into the body.",83357,27311,174660
+0,10,"Fitbit","Fitbit","Participant fitbit data",4,4,8460
+21,4,"Measurement","Labs & Measurements","Measurement",14034,13633,182000
+0,8,"Physical Measurement","Physical Measurements","Participants have the option to provide a standard set of physical measurements as part of the enrollment process",8,8,269780
+10,6,"Procedure","Procedures","Procedure",29712,27702,182000
