+++ public-api/db-cdr/csv/domain_vocabulary_info.csv
@@ -1,0 +1,36 @@
+"domain_id","vocabulary_id","all_concept_count","standard_concept_count"
+Condition,ICD10CM,26397,0
+Condition,ICD9CM,13917,0
+Condition,SNOMED,23599,45336
+Device,HCPCS,27,50
+Device,SNOMED,8,16
+Drug,CPT4,78,134
+Drug,HCPCS,663,36
+Drug,RxNorm,3752,7454
+Ethnicity,Ethnicity,2,4
+Gender,Gender,2,4
+Gender,SNOMED,3,0
+Meas Value,SNOMED,8,16
+Measurement,CPT4,952,1838
+Measurement,HCPCS,34,24
+Measurement,ICD9CM,27,0
+Measurement,LOINC,1674,3314
+Measurement,PPI,47,14
+Measurement,SNOMED,33,56
+Metadata,Domain,4,0
+Metadata,SNOMED,1,0
+Observation,CPT4,68,136
+Observation,HCPCS,141,242
+Observation,ICD10CM,238,0
+Observation,ICD9CM,1242,0
+Observation,PPI,194,352
+Observation,SNOMED,2834,5566
+Procedure,CPT4,7290,13202
+Procedure,HCPCS,258,234
+Procedure,ICD10PCS,5799,11598
+Procedure,ICD9CM,305,0
+Procedure,ICD9Proc,3125,6240
+Procedure,SNOMED,243,484
+Race,Race,4,8
+Spec Anatomic Site,SNOMED,12,22
+Visit,Visit,3,6
