+++ public-api/db-cdr/csv/survey_module.csv
@@ -1,0 +1,8 @@
+"concept_id","name","description","question_count","participant_count","order_number","can_show"
+1585855,"Lifestyle","Survey includes information on participant smoking, alcohol and recreational drug use.",26,218500,3,1
+1585710,"Overall Health","Survey includes information about how participants report levels of individual health.",21,219800,2,1
+1586134,"The Basics","This survey includes participant demographic information.",16,225140,1,1
+43529712,"Personal Medical History","This survey includes information about past medical history, including medical conditions and approximate age of diagnosis.",465,39320,4,1
+43528895,"Health Care Access & Utilization","Survey includes information about a participant’s access to and use of health care.",57,45820,5,1
+43528698,"Family Medical History","Survey includes information about the medical history of a participant’s immediate biological family members.",67,41600,6,1
+1333342,"COVID-19 Participant Experience","Survey includes information about the impact of COVID-19 on participant mental and physical health.",107,0,7,0
