+++ public-api/src/main/java/org/pmiops/workbench/cdr/dao/AchillesResultDistDao.java
@@ -1,18 +1,18 @@
 package org.pmiops.workbench.cdr.dao;
-import org.pmiops.workbench.cdr.model.AchillesResultDist;
+import org.pmiops.workbench.cdr.model.DbAchillesResultDist;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AchillesResultDistDao extends CrudRepository<AchillesResultDist, Long> {
-    List<AchillesResultDist> findAll();
+public interface AchillesResultDistDao extends CrudRepository<DbAchillesResultDist, Long> {
+    List<DbAchillesResultDist> findAll();
 
     @Query(value = "select * from achilles_results_dist where analysis_id=?1 and stratum_1=?2",nativeQuery=true)
-    List<AchillesResultDist> fetchConceptDistResults(Long analysisId, String conceptId);
+    List<DbAchillesResultDist> fetchConceptDistResults(Long analysisId, String conceptId);
 
     @Query(value = "select * from achilles_results_dist where analysis_id in (?1) and stratum_1 in (?2)", nativeQuery = true)
-    List<AchillesResultDist> fetchByAnalysisIdsAndConceptIds(List<Long> analysisIds,List<String> conceptIds);
+    List<DbAchillesResultDist> fetchByAnalysisIdsAndConceptIds(List<Long> analysisIds,List<String> conceptIds);
 
 }
